feat: Extract InfoPanel and SidebarTabPanel components, add sidebar config UI
Extract the info panel and sidebar tab logic from the monolithic reader
route into dedicated components. Add a settings panel for configuring
which tabs appear on each sidebar side with reordering support.
