+++ apps/desktop/src/locales/es.json
@@ -390,7 +390,8 @@
 		"citationPreview": "Vista previa de cita",
 		"cached": "En caché",
 		"fetched": "Obtenido",
-		"browser": "Navegador"
+		"browser": "Navegador",
+		"annotation": "Anotación"
 	},
 	"agent": {
 		"showHistory": "Mostrar historial",
@@ -577,6 +578,14 @@
 		"bilingualLayoutSideBySide": "Lado a lado (izquierda-derecha)",
 		"resetColumnLayout": "Restablecer disposicion de columnas",
 		"resetColumnLayoutNote": "Restablece el ancho, visibilidad y orden de las columnas a los valores predeterminados.",
+		"readerPanels": "Paneles del lector",
+		"readerPanelsDesc": "Configurar las pestañas de las barras laterales",
+		"leftSidebar": "Barra lateral izquierda",
+		"rightSidebar": "Barra lateral derecha",
+		"addTab": "Añadir pestaña...",
+		"moveUp": "Subir",
+		"moveDown": "Bajar",
+		"removeTab": "Eliminar",
 		"themeLight": "Claro",
 		"themeDark": "Oscuro",
 		"themeSystem": "Sistema",
